+++ src/signature-functions/core/signatures.ts
@@ -1,3 +1,8 @@
+import { action, context, github, storage } from "../../utils.ts";
+import { options } from "../options.ts";
+import { applicationType, storageVersion } from "../meta.ts";
+
+import type { SignatureStorage } from "./types.ts";
 import type { GitActor, SignatureData, SignatureStatus } from "./types.ts";
 
 /** Filter committers with their signature status */
@@ -25,4 +30,53 @@ export function getSignatureStatus(
     }
   }
   return status;
+}
+
+export const defaultSignatureContent: SignatureStorage = {
+  type: applicationType,
+  version: storageVersion,
+  data: {
+    documentSHA: "", // TODO
+    signatures: [],
+    superseded: [],
+    invalidated: [],
+  },
+};
+
+export type SignatureContent = github.Content<SignatureStorage>;
+
+export async function readSignatureStorage(): Promise<SignatureContent> {
+  switch (options.storage.type) {
+    case "local":
+    case "remote": {
+      const { content, sha } = await storage.readGithub(
+        options.storage,
+        JSON.stringify(defaultSignatureContent),
+        options.message.commit.setup,
+      );
+      return { content: JSON.parse(content), sha };
+    }
+    default:
+      action.fail("Unknown storage type");
+  }
+}
+
+export function writeSignatureStorage(content: SignatureContent) {
+  switch (options.storage.type) {
+    case "local":
+    case "remote":
+      return storage.writeGithub(
+        {
+          content: JSON.stringify(content.content),
+          sha: content.sha,
+        },
+        options.storage,
+        `${
+          options.message.commit.signed
+            .replace("${signatory}", context.payload.issue!.user.login)
+        }. Closes #${context.issue.number}`,
+      );
+    default:
+      action.fail("Unknown storage type");
+  }
 }
